Move test plugins to separate file
